+++ CHANGELOG.md
@@ -4,8 +4,23 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2026-09-22
+
+### Changed
+
+- doxyfile.meta ignore rule removed
+
+### Added
+
+- HiddenStreamingAssets~ folders and supporting Editor scripts
+
+### Removed
+
+- streamingassets.unitypackage
+- Tests (empty)
+
 ## [0.1.0] - 2026-07-17
 
 ### This is the first release of *\<RLMG Tools - Media Players\>*.
 
-*Short description of this release*
+Initial commit with image sequence and video playback and loading, with samples.